+++ mediation/Chartboost/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Chartboost Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
+## 1.1.0
+- Supports Chartboost Android SDK version 7.2.0.
+- Supports Chartboost iOS SDK version 7.2.0.
+- Added `Chartboost.RestrictDataCollection()` method to enable GDPR data collection restrictions for the Chartboost SDK.
+
 ## 1.0.0
 - First release!
 - Supports Chartboost Android SDK version 7.0.1.